fix: solve conflicts
Use SingleApplication to to check if any of the processes are already running, and if so, causes the original process to display the main window.
Add multi-theme function.
Allow color themes to change eagerly.
When the system color theme switches, the software also switches.
enhance the logic of add and remove ListItemWidget
Create Basic UI of FileInformationWidget